+++ src/ops_data_store/config.py
@@ -10,6 +10,10 @@ class Config:
         """Create Config instance and load options from dotenv file."""
         load_dotenv()
 
+    def dump(self) -> dict:
+        """Return application configuration as a dictionary."""
+        return {"VERSION": self.VERSION}
+
     @property
     def VERSION(self) -> str:
         """Application version."""