+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 
 * note in release issue template to add link to README in each release
   [#15](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/15)
+* `config show` CLI command
+  [#18](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/18)
 
 ### Changed